% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/answers.R
\docType{data}
\name{answers}
\alias{answers}
\title{A dataset of David Robinson's Stack Overflow answers as of 4/1/2015}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	742 obs. of  14 variables:
 $ owner_reputation   : int  34979 34979 34979 34979 34979 34979 34979 34979 34979 34979 ...
 $ owner_user_id      : int  712603 712603 712603 712603 712603 712603 712603 712603 712603 712603 ...
 $ owner_user_type    : chr  "registered" "registered" "registered" "registered" ...
 $ owner_accept_rate  : int  100 100 100 100 100 100 100 100 100 100 ...
 $ owner_profile_image: chr  "https://www.gravatar.com/avatar/69c6cc9d66b7c8a622776afcc338619f?s=128&d=identicon&r=PG" "https://www.gravatar.com/avatar/69c6cc9d66b7c8a622776afcc338619f?s=128&d=identicon&r=PG" "https://www.gravatar.com/avatar/69c6cc9d66b7c8a622776afcc338619f?s=128&d=identicon&r=PG" "https://www.gravatar.com/avatar/69c6cc9d66b7c8a622776afcc338619f?s=128&d=identicon&r=PG" ...
 $ owner_display_name : chr  "David Robinson" "David Robinson" "David Robinson" "David Robinson" ...
 $ owner_link         : chr  "http://stackoverflow.com/users/712603/david-robinson" "http://stackoverflow.com/users/712603/david-robinson" "http://stackoverflow.com/users/712603/david-robinson" "http://stackoverflow.com/users/712603/david-robinson" ...
 $ is_accepted        : logi  FALSE FALSE FALSE TRUE TRUE TRUE ...
 $ score              : int  0 1 1 1 4 1 3 1 1 1 ...
 $ last_activity_date : POSIXct, format: "2015-03-31 13:23:02" "2015-03-31 10:12:57" ...
 $ creation_date      : POSIXct, format: "2015-03-31 13:23:02" "2015-03-31 10:12:57" ...
 $ answer_id          : int  29376812 29373378 29269729 28948541 28947158 28947096 28737626 28703683 28703301 28702838 ...
 $ question_id        : int  29376369 29371325 29269708 28948168 28947043 28411989 28735176 28703583 28702974 28702761 ...
 $ last_edit_date     : POSIXct, format: NA NA ...
 - attr(*, "metadata")=List of 3
  ..$ has_more       : logi FALSE
  ..$ quota_max      : int 10000
  ..$ quota_remaining: int 9992
}}
\usage{
answers
}
\description{
This object was created with the lines of code shown in the Examples
below. It is a dataset from the Stack Exchange API showing answers from
user 712603 (David Robinson, this package's maintainer). It is useful for
example plots, such as those in the vignette, since everyone knows
freehand red circles are most important when plotting Stack Overflow data.
}
\examples{
# created with stackr:
\dontrun{
    require("stackr")
    answers <- stack_users(712603, "answers", num_pages = 10, pagesize = 100)
    answers <- dplyr::tbl_df(answers)
}
}
\keyword{datasets}

