% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom-freehand.R
\name{geom_freehand}
\alias{geom_freehand}
\title{Freehand circles}
\usage{
geom_freehand(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
  include aesthetics whose values you want to set, not map. See
  \code{\link{layer}} for more details.}
}
\description{
Add freehand circles (default red) to a plot, similar to the style of Meta
Stack Overflow (http://meta.stackexchange.com/a/19775/176330).
}
\details{
The \code{radius} argument determines the radius of the freehand circles. It
is normalized based on the scales such that a radius of 1 takes up 1/25 of the
plot's width and height. You can also set the radius separately for the x- and
y- axes with \code{radiusx} and \code{radiusy}.
}
\section{Aesthetics}{

\code{geom_freehand} understands the following aesthetics (required aesthetics are in bold):

\itemize{
  \item \code{\strong{x}}
  \item \code{\strong{y}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{linetype}
  \item \code{noisiness}
  \item \code{radius}
  \item \code{radiusx}
  \item \code{radiusy}
  \item \code{size}
}
}
\examples{
require("ggplot2")

p <- ggplot(mtcars, aes(wt, mpg))
p + geom_point()

one_car <- mtcars[1, ]
p + geom_point() + geom_freehand(data = one_car, rx = .1, ry = .5)
}

